requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-telegram-bot==20.3
pytz==2023.3
requests==2.29.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
urllib3==1.26.15
